requirements.txt
pymongo[srv]==3.12
pandas
numpy
scikit-learn
python-dotenv


# -e .  #editable mode